+++ mmdet/core/anchor/__init__.py
@@ -1,8 +1,8 @@
 from .anchor_generator import AnchorGenerator
 from .point_generator import PointGenerator
-from .utils import anchor_inside_flags, calc_region, images_to_levels, unmap
+from .utils import anchor_inside_flags, calc_region, images_to_levels
 
 __all__ = [
     'AnchorGenerator', 'anchor_inside_flags', 'PointGenerator',
-    'images_to_levels', 'unmap', 'calc_region'
+    'images_to_levels', 'calc_region'
 ]
