+++ mmdet/core/utils/__init__.py
@@ -1,6 +1,7 @@
 from .dist_utils import DistOptimizerHook, allreduce_grads
-from .misc import multi_apply, tensor2imgs
+from .misc import multi_apply, tensor2imgs, unmap
 
 __all__ = [
-    'allreduce_grads', 'DistOptimizerHook', 'tensor2imgs', 'multi_apply'
+    'allreduce_grads', 'DistOptimizerHook', 'tensor2imgs', 'multi_apply',
+    'unmap'
 ]
